(#32) Refactoring all existing formulas
