+++ blueprints/codes/eurocode/nen_en_1992_1_1_c2_2011/chapter_3_materials/__init__.py
@@ -1,0 +1,1 @@
+"""Module containing all formulas from NEN-EN 1992-1-1+C2:2011: Chapter 3 - Materials."""
